requirements.txt
pip
websocket-client~=0.57.0
prawcore~=1.3.0
requests~=2.7.0
setuptools~=41.2.0
praw~=7.0.0
python-dotenv~=0.13.0
pytelegrambotapi